+++ src/Interfaces/DocumentInterface.php
@@ -2,15 +2,19 @@
 
 namespace Swis\JsonApi\Client\Interfaces;
 
+use Psr\Http\Message\ResponseInterface;
 use Swis\JsonApi\Client\Collection;
 use Swis\JsonApi\Client\ErrorCollection;
+use Swis\JsonApi\Client\Jsonapi;
+use Swis\JsonApi\Client\Links;
+use Swis\JsonApi\Client\Meta;
 
 interface DocumentInterface extends \JsonSerializable
 {
     /**
      * @return \Psr\Http\Message\ResponseInterface|null
      */
-    public function getResponse();
+    public function getResponse(): ? ResponseInterface;
 
     /**
      * @return \Swis\JsonApi\Client\Interfaces\DataInterface
@@ -30,12 +34,12 @@ interface DocumentInterface extends \JsonSerializable
     /**
      * @return \Swis\JsonApi\Client\Meta|null
      */
-    public function getMeta();
+    public function getMeta(): ? Meta;
 
     /**
      * @return \Swis\JsonApi\Client\Links|null
      */
-    public function getLinks();
+    public function getLinks(): ? Links;
 
     /**
      * @return mixed
@@ -45,7 +49,7 @@ interface DocumentInterface extends \JsonSerializable
     /**
      * @return \Swis\JsonApi\Client\Jsonapi|null
      */
-    public function getJsonapi();
+    public function getJsonapi(): ? Jsonapi;
 
     /**
      * @return array