+++ src/Interfaces/ItemInterface.php
@@ -51,7 +51,7 @@ interface ItemInterface extends DataInterface
     /**
      * @return \Swis\JsonApi\Client\Links|null
      */
-    public function getLinks();
+    public function getLinks(): ? Links;
 
     /**
      * @param \Swis\JsonApi\Client\Meta|null $meta
@@ -63,7 +63,7 @@ interface ItemInterface extends DataInterface
     /**
      * @return \Swis\JsonApi\Client\Meta|null
      */
-    public function getMeta();
+    public function getMeta(): ? Meta;
 
     /**
      * @param array $attributes