+++ src/Interfaces/ManyRelationInterface.php
@@ -50,7 +50,7 @@ interface ManyRelationInterface
     /**
      * @return \Swis\JsonApi\Client\Links|null
      */
-    public function getLinks();
+    public function getLinks(): ? Links;
 
     /**
      * @param \Swis\JsonApi\Client\Meta|null $meta
@@ -60,5 +60,5 @@ interface ManyRelationInterface
     /**
      * @return \Swis\JsonApi\Client\Meta|null
      */
-    public function getMeta();
+    public function getMeta(): ? Meta;
 }