+++ src/Interfaces/TypeMapperInterface.php
@@ -11,19 +11,19 @@ interface TypeMapperInterface
      * @param string $type
      * @param string $class
      */
-    public function setMapping(string $type, string $class);
+    public function setMapping(string $type, string $class): void;
 
     /**
      * @param string $type
      *
      * @return bool
      */
-    public function hasMapping(string $type);
+    public function hasMapping(string $type): bool;
 
     /**
      * @param string $type
      */
-    public function removeMapping(string $type);
+    public function removeMapping(string $type): void;
 
     /**
      * @param string $type
@@ -32,5 +32,5 @@ interface TypeMapperInterface
      *
      * @return \Swis\JsonApi\Client\Interfaces\ItemInterface
      */
-    public function getMapping(string $type);
+    public function getMapping(string $type): ItemInterface;
 }