+++ src/Traits/HasMeta.php
@@ -26,7 +26,7 @@ trait HasMeta
     /**
      * @return \Swis\JsonApi\Client\Meta|null
      */
-    public function getMeta()
+    public function getMeta(): ? Meta
     {
         return $this->meta;
     }